Final touches. Change bindings to gp-bindings
