+++ dggrs/src/adapters/dggrid/isea3h.rs
@@ -10,7 +10,7 @@
 use crate::adapters::dggrid::common;
 use crate::adapters::dggrid::dggrid::DggridAdapter;
 use crate::error::port::PortError;
-use crate::models::common::{Zone, Zones};
+use crate::models::common::Zones;
 use crate::ports::dggrs::DggrsPort;
 use core::f64;
 use geo::geometry::Point;
@@ -19,7 +19,6 @@ use std::fs::OpenOptions;
 use std::io::{self, Write};
 use std::path::PathBuf;
 use tracing::debug;
-use wasm_bindgen::JsValue;
 pub const CLIP_CELL_DENSIFICATION: u8 = 50; // DGGRID option
 
 pub struct Isea3hImpl {
@@ -44,66 +43,6 @@ impl Default for Isea3hImpl {
 }
 
 impl DggrsPort for Isea3hImpl {
-    // fn zones_from_bbox1(
-    //     &self,
-    //     depth: u8,
-    //     densify: bool,
-    //     bbox: Option<Vec<Vec<f64>>>,
-    // ) -> Result<Zones, PortError> {
-    //     // THIS WORKS
-    //     let (meta_path, aigen_path, children_path, neighbor_path, bbox_path, _input_path) =
-    //         common::dggrid_setup(&self.adapter.workdir);
-
-    //     let _ = common::dggrid_metafile(
-    //         &meta_path,
-    //         &depth,
-    //         &aigen_path.with_extension(""),
-    //         &children_path.with_extension(""),
-    //         &neighbor_path.with_extension(""),
-    //         densify,
-    //     );
-
-    //     let _ = isea3h_metafile(&meta_path);
-
-    //     // if let Some(bbox) = &bbox {
-    //     //     let _ = common::bbox_to_aigen(bbox, &bbox_path);
-
-    //     //     // Append to metafile
-    //     //     // let mut meta_file = OpenOptions::new()
-    //     //     //     .append(true)
-    //     //     //     .write(true)
-    //     //     //     .open(&meta_path)
-    //     //     //     .expect("cannot open file");
-    //     //     // append_to_file(meta_path, &contents);
-    //     //     let contents = format!(
-    //     //         "clip_subset_type AIGEN\n\
-    //     //  clip_region_files {} \n\
-    //     //  ",
-    //     //         &bbox_path.to_string_lossy()
-    //     //     );
-
-    //     //     let _ = append_to_file(&meta_path, &contents);
-    //     //     // let _ = writeln!(meta_file, "clip_subset_type AIGEN");
-    //     //     // let _ = writeln!(
-    //     //     //     meta_file,
-    //     //     //     "clip_region_files {}",
-    //     //     //     &bbox_path.to_string_lossy()
-    //     //     // );
-    //     // }
-
-    //     common::print_file(meta_path.clone());
-    //     common::dggrid_execute(&self.adapter.executable, &meta_path);
-    //     let result = common::dggrid_parse(&aigen_path, &children_path, &neighbor_path, &depth)?;
-
-    //     common::dggrid_cleanup(
-    //         &meta_path,
-    //         &aigen_path,
-    //         &children_path,
-    //         &neighbor_path,
-    //         &bbox_path,
-    //     );
-    //     Ok(result)
-    // }
     fn zones_from_bbox(
         &self,
         depth: u8,
@@ -325,54 +264,17 @@ impl DggrsPort for Isea3hImpl {
 pub fn isea3h_metafile(meta_path: &PathBuf) -> io::Result<()> {
     debug!("Writing to {:?}", meta_path);
     // Append to metafile format
-    // let mut meta_file = OpenOptions::new()
-    //     .append(true)
-    //     .write(true)
-    //     .open(&meta_path)
-    //     .expect("cannot open file");
-    // writeln!(meta_file, "dggs_type {}", "ISEA3H")?;
-    // writeln!(meta_file, "dggs_aperture 3")?;
-    // writeln!(meta_file, "output_address_type Z3")?;
-    let contents = format!(
-        "dggs_type {}\n\
-         dggs_aperture 3\n\
-         output_address_type Z3\n",
-        "ISEA3H"
-    );
-
-    let _ = append_to_file(meta_path, &contents);
-    // .map_err(|e| format!("Failed to write metafile: {}", e));
-    Ok(())
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-pub fn append_to_file(path: &std::path::Path, contents: &str) -> std::io::Result<()> {
-    let mut f = OpenOptions::new()
+    let mut meta_file = OpenOptions::new()
         .append(true)
         .write(true)
-        .create(true)
-        .open(path)?;
-    write!(f, "{}", contents)?;
+        .open(&meta_path)
+        .expect("cannot open file");
+    writeln!(meta_file, "dggs_type {}", "ISEA3H")?;
+    writeln!(meta_file, "dggs_aperture 3")?;
+    writeln!(meta_file, "output_address_type Z3")?;
 
     Ok(())
 }
-
-#[cfg(target_arch = "wasm32")]
-pub fn append_to_file(path: &std::path::Path, contents: &str) -> Result<(), JsValue> {
-    //     let contents = format!(
-    //         "dggs_type {}
-    // \
-    //          dggs_aperture 3
-    // \
-    //          output_address_type Z3
-    // ",
-    //         "ISEA3H"
-    //     );
-
-use crate::append_to_file_wasm;
-    append_to_file_wasm(&path.to_string_lossy(), &contents)
-}
-
 
 pub fn extract_res_from_cellid(id: &str, dggs_type: &str) -> Result<u8, String> {
     match dggs_type {
