+++ dggrs/src/adapters/h3o/h3.rs
@@ -9,7 +9,6 @@
 
 use crate::adapters::h3o::common::{cells_to_zones, res};
 use crate::adapters::h3o::h3o::H3oAdapter;
-use crate::error::dggrid::DggridError;
 use crate::error::h3o::H3oError;
 use crate::error::port::PortError;
 use crate::models::common::Zones;
@@ -17,8 +16,6 @@ use crate::ports::dggrs::DggrsPort;
 use geo::{LineString, Point, Polygon};
 use h3o::geom::{ContainmentMode, TilerBuilder};
 use h3o::{CellIndex, LatLng, Resolution};
-use wasm_bindgen::JsValue;
-use std::path::PathBuf;
 use std::str::FromStr;
 
 pub const MAX_DEPTH: u8 = 10;
