+++ gp-bindings/nodejs/.gitignore
@@ -2,3 +2,6 @@ node_modules
 dist
 *.log
 .DS_Store
+.vscode
+.github
+.pnp-lock.yaml
