+++ gp-bindings/nodejs/src/index.ts
